Now have a large pdf file and a copy of the discussion
The pdf came from doxygen and MikeTex, one latex application.  It
includes a lot of information, possibly more than Dalibor needs...but it
is something and one particular part includes a lot of the comments that
are required.  I have also added the latest version of source code and
the discussion as written so far.